change(inclusion request): add the user during transform
# Conflicts:
#	cds_migrator_kit/rdm/records/load/load.py
