+++ cds_migrator_kit/rdm/records/load/load.py
@@ -48,61 +48,6 @@ from cds_migrator_kit.errors import (
 )
 
 
-def _is_email(value):
-    """Return True if the reviewer value looks like an email address."""
-    return "@" in value
-
-
-def _parse_reviewer_name(name):
-    """Split a 'Family, Given' or 'Given Family' string into (family, given).
-
-    ``request_reviewers`` (906__p) stores names as "Given Family" (comma
-    already resolved), but legacy data can also arrive as "Family, Given".
-    """
-    name = name.strip()
-    if "," in name:
-        family, _, given = name.partition(",")
-        return family.strip(), given.strip()
-    parts = name.split()
-    if len(parts) > 1:
-        return parts[-1], " ".join(parts[:-1])
-    return name, ""
-
-
-def find_reviewer(reviewer):
-    """Resolve a reviewer string (email or name) to a User.
-
-    :param reviewer: email address, or a "Family, Given"/"Given Family" name.
-    :raises RecordFlaggedCuration: if no matching user is found, so the
-        record is flagged for manual curation instead of failing outright.
-    """
-    reviewer = reviewer.strip()
-    if _is_email(reviewer):
-        user = User.query.filter_by(email=reviewer).one_or_none()
-    else:
-        family_name, given_name = _parse_reviewer_name(reviewer)
-        query = User.query.filter(
-            db.func.lower(User._user_profile["family_name"].as_string())
-            == family_name.lower()
-        )
-        if given_name:
-            query = query.filter(
-                db.func.lower(User._user_profile["given_name"].as_string())
-                == given_name.lower()
-            )
-        user = query.one_or_none()
-
-    if user is None:
-        raise RecordFlaggedCuration(
-            message=f"Reviewer '{reviewer}' could not be matched to an account.",
-            field="request_reviewers",
-            stage="load",
-            value=reviewer,
-        )
-
-    return user
-
-
 def import_legacy_files(filepath):
     """Download file from legacy."""
     if current_app.config["CDS_MIGRATOR_KIT_ENV"] == "local":
@@ -121,7 +66,6 @@ class CDSRecordServiceLoad(Load):
         self,
         db_uri=None,
         data_dir=None,
-        tmp_dir=None,
         entries=None,
         dry_run=False,
         legacy_pids_to_redirect=None,
@@ -484,19 +428,7 @@ class CDSRecordServiceLoad(Load):
             return
 
         status = request_data.get("status", "accepted")
-        reviewer_names = request_data.get("reviewers", [])
-        reviewers = []
-        for name in reviewer_names:
-            try:
-                reviewers.append(find_reviewer(name))
-            except RecordFlaggedCuration as exc:
-                self.migration_logger.add_information(
-                    legacy_recid,
-                    {"message": exc.message, "value": exc.value},
-                )
-                # keep a placeholder so the reviewer count/order is preserved;
-                # resolved to the "-1" sentinel user id below
-                reviewers.append(None)
+        reviewers = request_data.get("reviewers", [])
 
         created_at = datetime.datetime.fromisoformat(record["created"])
 
@@ -537,8 +469,7 @@ class CDSRecordServiceLoad(Load):
         request.model.created = created_at
 
         if reviewers:
-            reviewers_payload = [{"user": str(r.id) if r else "-1"} for r in reviewers]
-            request.reviewers = reviewers_payload
+            request.reviewers = reviewers
 
         if status:
             request.status = status
